+++ getPEM.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+if [ "$NETLIFY" = true ] ; then
+    echo 'Build On Netlify, writing PEM File'
+	printenv GITHUB_PRIVATEKEY | base64 -d -o github.pem
+fi
