+++ src/utils/config.ts
@@ -1,7 +1,7 @@
-function base64(data: string) {
-  const buff = Buffer.from(data, 'base64')
-  return buff.toString('utf8')
-}
+// function base64(data: string) {
+//   const buff = Buffer.from(data, 'base64')
+//   return buff.toString('utf8')
+// }
 
 const getEnvironmentVariable = (environmentVariable: string): string => {
   const unvalidatedEnvironmentVariable = process.env[environmentVariable]
@@ -20,8 +20,6 @@ const getEnvironmentVariable = (environmentVariable: string): string => {
 
 export const config = {
   GITHUB_APPID: getEnvironmentVariable('GITHUB_APPID'),
-  GITHUB_PRIVATEKEY: process.env.NETLIFY
-    ? base64(getEnvironmentVariable('GITHUB_PRIVATEKEY')).replace(/\\n/g, '\n')
-    : getEnvironmentVariable('GITHUB_PRIVATEKEY'),
+  GITHUB_PRIVATEKEY: getEnvironmentVariable('GITHUB_PRIVATEKEY'),
   GITHUB_INSTALLATIONID: getEnvironmentVariable('GITHUB_INSTALLATIONID'),
 }
