+++ src/utils/getDocsList.preval.ts
@@ -4,36 +4,8 @@ const docsList = {
   'billing-options': 'docs/guides/Getting Started/catalog-overview.md',
 }
 
-import { Octokit } from 'octokit'
-import { createAppAuth } from '@octokit/auth-app'
-import { throttling } from '@octokit/plugin-throttling'
+import octokit from 'utils/octokitConfig'
 
-const MyOctokit = Octokit.plugin(throttling)
-
-const octokit = new MyOctokit({
-  authStrategy: createAppAuth,
-  auth: {
-    appId: process.env.GITHUB_APPID,
-    privateKey: process.env.GITHUB_PRIVATEKEY,
-    installationId: process.env.GITHUB_INSTALLATIONID,
-  },
-  throttle: {
-    onRateLimit: (retryAfter: any, options: any, octokit: any) => {
-      octokit.log.warn(
-        `Request quota exhausted for request ${options.method} ${options.url}`
-      )
-      octokit.log.info(`Retrying after ${retryAfter} seconds!`)
-      return true
-    },
-    onSecondaryRateLimit: (retryAfter: any, options: any, octokit: any) => {
-      retryAfter = retryAfter
-      // does not retry, only logs a warning
-      octokit.log.warn(
-        `SecondaryRateLimit detected for request ${options.method} ${options.url}`
-      )
-    },
-  },
-})
 async function getGithubTree(org: string, repo: string, ref: string) {
   const response = octokit.request(
     'GET /repos/{org}/{repo}/git/trees/{ref}?recursive=true',
