+++ src/utils/getDocsList.ts
@@ -3,43 +3,8 @@ const docsList = {
   'billing-options': 'docs/guides/Getting Started/catalog-overview.md',
 }
 
-import { Octokit } from 'octokit'
-import { createAppAuth } from '@octokit/auth-app'
-import { throttling } from '@octokit/plugin-throttling'
-import { config } from 'utils/config'
+import octokit from 'utils/octokitConfig'
 
-const MyOctokit = Octokit.plugin(throttling)
-
-const octokitConfig = {
-  authStrategy: createAppAuth,
-  auth: {
-    appId: config.GITHUB_APPID,
-    privateKey: config.GITHUB_PRIVATEKEY,
-    installationId: config.GITHUB_INSTALLATIONID,
-  },
-  throttle: {
-    onRateLimit: (retryAfter: any, options: any, octokit: any) => {
-      octokit.log.warn(
-        `Request quota exhausted for request ${options.method} ${options.url}`
-      )
-      octokit.log.info(`Retrying after ${retryAfter} seconds!`)
-      return true
-    },
-    onSecondaryRateLimit: (retryAfter: any, options: any, octokit: any) => {
-      retryAfter = retryAfter
-      // does not retry, only logs a warning
-      octokit.log.warn(
-        `SecondaryRateLimit detected for request ${options.method} ${options.url}`
-      )
-    },
-  },
-}
-
-if (process.env.NETLIFY) {
-  console.log(JSON.stringify(octokitConfig.auth))
-}
-
-const octokit = new MyOctokit(octokitConfig)
 async function getGithubTree(org: string, repo: string, ref: string) {
   const response = octokit.request(
     'GET /repos/{org}/{repo}/git/trees/{ref}?recursive=true',
