+++ src/utils/getGithubFile.ts
@@ -1,41 +1,4 @@
-/* eslint-disable @typescript-eslint/no-explicit-any, @typescript-eslint/ban-ts-comment */
-import { Octokit } from 'octokit'
-import { createAppAuth } from '@octokit/auth-app'
-import { throttling } from '@octokit/plugin-throttling'
-import { config } from 'utils/config'
-
-const MyOctokit = Octokit.plugin(throttling)
-
-const octokitConfig = {
-  authStrategy: createAppAuth,
-  auth: {
-    appId: config.GITHUB_APPID,
-    privateKey: config.GITHUB_PRIVATEKEY,
-    installationId: config.GITHUB_INSTALLATIONID,
-  },
-  throttle: {
-    onRateLimit: (retryAfter: any, options: any, octokit: any) => {
-      octokit.log.warn(
-        `Request quota exhausted for request ${options.method} ${options.url}`
-      )
-      octokit.log.info(`Retrying after ${retryAfter} seconds!`)
-      return true
-    },
-    onSecondaryRateLimit: (retryAfter: any, options: any, octokit: any) => {
-      retryAfter = retryAfter
-      // does not retry, only logs a warning
-      octokit.log.warn(
-        `SecondaryRateLimit detected for request ${options.method} ${options.url}`
-      )
-    },
-  },
-}
-
-if (process.env.NETLIFY) {
-  console.log(JSON.stringify(octokitConfig.auth))
-}
-
-const octokit = new MyOctokit(octokitConfig)
+import octokit from 'utils/octokitConfig'
 
 export default async function getGithubFile(
   owner: string,
@@ -52,7 +15,7 @@ export default async function getGithubFile(
     mediaType: {
       format: 'raw',
     },
-  })
-  // @ts-ignore
-  return response.data
+  }) // eslint-disable-line
+  // @ts-ignore // eslint-disable-line
+  return response.data // eslint-disable-line
 }
